arsenal/requirements.txt
numpy==2.1.1
cython===3.0.11
pyqlib==0.9.5.99